backend/requirements.txt
Flask
flask_cors
Gunicorn
pymongo[srv]
Flask-PyMongo
python-dotenv
flask-jwt-extended